Going back to the abstract
The abstract is the most important part of the paper, and I think it needs some rewriting. Please try and read it like someone that is not familiar either with the problem or the methodology. The objectives and motivation of the paper must be very clear, as well ad the results that have been obtained. CC @amherag @mariosky